rename services to service everywhere
