+++ src/handlers/mod.rs
@@ -1,5 +1,5 @@
 pub mod auth;
 pub mod image;
-pub mod services;
+pub mod service;
 pub mod user;
 pub mod version;
